feat: add config
Adds a config package, with a struct representing the application
configuration. This allows defining logging configuration and some
properties for the api server.

Also adds a test script to be used for running tests, and generating
coverage.

Adds a debug command, which has a show-config subcommand, mostly
useful for testing and seeing the affect of environment variable
and cli flags have on the config used in the app.

Updates the server to use the port as defined in the config rather
than being hardcoded.