fix: github oauth token endpoint
Fixes the github oauuth token endpoint. The main issue was actually
the config; it was not properly parsing the yaml due to a conflict
between the yaml and mapstructure tags in the struct.

Adds a logger to the github oauth client so that we can throw some
debug logs in there.