refactor: fix user repo, and add tests
Fixes the issue in the user repo where not finding any results was
still returning an error. This was due to looking for the wrong type
of error from the sql package rather than pgx.

Adds a whole load for setting up testcontainers. This is primarily
used to spin up a postgres (more specifically timescale DB) for any
tests in the postgres package. This covers running all the migrations
for each test, and creating a separate schema for each test. This
allows tests to run in parallel without interfering with eachother
but doesn't need a separate container for each test.

Fixed an issue with hydrating user models, it just needed to parse
the string for the UUID rather than casting it to bytes.