refactor: add save method to user repository
Adds a save method to the user repository. Uses an upsert for both
the user and github_user table (only a single method for saving a
user exists as it's an aggregate root). Tests cover some scenarios
around validation and upsertion.

Updated migrations to include unique constraints where appropriate
to work with the upsert commands.