refactor: setup user creation on registration
Fills in logic of the auth login endpoint so that once the access
token is verified with github we fetch the users name, email, and
node id from github and create them. If the user has previously
created an account we do nothing.

Puts in a placeholder for generating a JWT for our application upon
login with github.

Splits the user command and query handlers. Primarily as this allows
better interface segregation (the query bus can only use the read
methods on the repo, and the command bus only knows about the save
method).