feat: setup refresh flow for JWT's
Sets up the refresh API endpoint for JWT authentication. The new
endpoint does not have auth, as i think is expected; the auth
middleware right now validates the JWT, which may have expired while
the refresh token hasn't. Not sure about this, but will circle back
at some point.

Adds a new config parameter for an HMAC key which is used when
hashing the refresh tokens, before storing them in the DB. This
also includes a new script and orca extension that generates a
random secret key that can be used for secrets like this. There is
a single hasher implementation that uses HMAC and sha256 to hash
values before storing.

Adds a new refresh token repository to manage the lifecycle for the
refresh tokens.

Adds a new pkg/dto module to contain some useful data structures
that are used in infra and application layer. It feels a bit odd,
but I shouldn't really have either of those layers depending on
eachother, and the refresh tokens, and hashing are not really
much of  domain concern (at least right now, the HashedValue
arguably could be).

Unified the management of JWT's and refresh tokens under a single
token manager.