feat: add logout
Adds a logout endpoint that deletes the refresh token, and all the
cookies. This does mean that the auth_token is still valid for a
short while, but the refresh token won't work so it shouldn't be
a massive problem. A reasonable trade-off to avoid hitting the DB
on every request.