fix: fix trivy vulnerabilities
Adds a user to the dockerfile, not absolutely necessary as this is
only for dev, but easier to do it and keep trivy happy, rather than
ignore and potentially skip this check later for production images.

Adds an orca extension to run trivy scan locally.

Adds a trivyignore file that prevents it failing because of a
missing healthcheck. I don't think it wants a healthcheck even in
the prod image, but that may change later. For now getting rid of
that check seems sensible.