refactor: add dev dockerfile and test workflow
Replaces the Dockerfile with a Dockerfile.dev. Main reason for this
is that the testcontainers require root privileges, but in order to
make trivy happy we shouldn't use the root user. Given we're only
using this file in dev it seems much easier to have this separate
dockerfile rather than having to build a whole load of nasty
workarounds for file permissions; considering how annoying this can
be to across different OS'.

This became apparrent when adding the test workflow, running them
locally had broken since updating the dockerfile.

Adds a workflow to run any tests in the repository.

Updates trivy config to ignore the new dockerfile, and updates
trivy ignore as we don't need to ignore anything right now.