refactor: various things (i know, bad commit)
Started by refactoring how we structure the layers in the application
mostly focused around how define the bounded contexts and what lives
in each. Mainly removed the top-level application package and moved
the logic that was associated with a specific domain into the
relevant context package. Reflected this for domain as well.

There were some things that are not actually relevant to a specific
context; most of this i'm associating with the presentation layer,
as it's all about api auth. This all lives in the api package.

Also looked into testing, and introduced some actual feature testing
although right now it just tests the startup probe. This gives a
basic foundation for adding more tests at a feature level.

Also moved the sqlc files closer to the repositories, hopefully
making it more obvious where the queries come from.