refactor: add actor middleware
Adds a middleware to load an actor based on the auth token. This
requires quite a few changes in a few places to make this work.
Adds new organisations and users bridhe to the top-level infra
package, which allow that to read data from other contexts.
Introduces an actor loader that runs as part of the auth middleware
to fetch basic data about who is performing an action. Adds some new
commands in organisations contexts to support this, they're needed
later anyway.