feat: support api key auth
Adds support for authenticating using an organisation level API
key; later this will be easy to extend for user level API keys.

This includes a bit of a refactor around the authentication middleware
allowing for us to check multiple authentication methods in a
priority order.

As part of this also improved the access logging. Firstly to ensure
that any request is logged, previously only requests that got past
all the middleware were logged. Secondly, includes meta data about
the auth method and api key used. Finally, includes the "operation
id" from the routes. This will allow for better aggregation/searching
in the access logs, where uri's may have query params that make it
hard to find all responses from the same endpoint. In order to do
this, we have to set some headers on the response as part of one
middlware, and then use those headers in the logging middleware to
include the relevant attributes.

Minor change for dev; removed the format on build as this was
causing some issues in the IDE as the file was often changed by
the formatter while I was also editing it, and vs code needed some
help to decide which change should take precendence.